Mover método 'exibirTodosOsCadastros' para classe ContaService
